Code moved out of UnityStudioForm and into new class UnityStudio.cs
