+++ Unity Studio/Unity Studio Classes/AssetPreloadData.cs
