+++ Unity Studio/Unity Studio Classes/AssetsFile.cs
@@ -453,6 +453,5 @@ namespace Unity_Studio
                 ClassStructures.Add(classID, aClass);
             }
         }
-
     }
 }
