+++ Unity Studio/Unity Studio Classes/BundleFile.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.IO;
-using SevenZip;
 using Lz4;
 
 namespace Unity_Studio
