+++ Unity Studio/Unity Studio Classes/EndianStream.cs
