+++ Unity Studio/Unity Studio Classes/helpers.cs
